Clean up images in IncorrectPermsSetModal
