+++ ui/src/pages/MyPost/IncorrectPermsSetModal/index.ts
@@ -1,0 +1,1 @@
+export { IncorrectPermsSetModal } from "./IncorrectPermsSetModal";
